backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

google-generativeai==0.3.1
openai==1.3.7

faiss-cpu==1.12.0
numpy==1.26.4

pydantic==2.5.0
pydantic-settings==2.1.0

python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

